% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_station}
\alias{check_station}
\title{Check if a station name produces a unique station ID.}
\usage{
check_station(station)
}
\arguments{
\item{station}{A station name to check for validity.}
}
\value{
A valid station ID, or an error if no or multiple stations match.
}
\description{
Check if a station name produces a unique station ID.
}
\keyword{internal}
