% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_url_response}
\alias{check_url_response}
\title{Check URL for invalid response}
\usage{
check_url_response(url_output)
}
\arguments{
\item{url_output}{The output from a URL call to check for invalid response}
}
\value{
If the response appears valid, return the original input, otherwise throw error.
}
\description{
Check URL for invalid response
}
\keyword{internal}
