% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathervane.R
\name{download_data}
\alias{download_data}
\title{Download the weather data from the constructed URL}
\usage{
download_data(url)
}
\arguments{
\item{url}{A character string containing the download URL}
}
\value{
A data.frame containing the specified weather data
}
\description{
Return a data frame containing the SILO weather dataset,
complete with prettied column names, from the given download
URL.
}
\examples{
weathervane:::download_data(
  weathervane:::build_url(
    -34.9, 138.6, start_date = '2020-01-01',
    finish_date = '2020-12-31', variables = c('rainfall')
  )
)
}
\seealso{
\code{\link[=build_url]{build_url()}} to generate the requisite URLs.
}
\keyword{internal}
