% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathervane.R
\name{download_url}
\alias{download_url}
\title{(Private) The download URL for the SILO weather data download}
\usage{
download_url(latitude, longitude, start_date, finish_date, variables)
}
\arguments{
\item{latitude}{The latitude (in decimal degrees North)}

\item{longitude}{The longitude (in decimal degrees East)}

\item{start_date}{A string or Date object for the starting date}

\item{finish_date}{A string or Date object for the finish date}

\item{variables}{A vector containing the variable names}
}
\value{
The parameter-formatted URL for the data download
}
\description{
Return a character string containing the parameter-formatted
download URL for the SILO data, given the latitude/longitude
coordinates, the start and finish dates of interest, and a list
of variables to be retrieved.
}
\examples{
weathervane:::download_url(-34.9, 138.6, '2020-01-01', '2020-12-31', c('rainfall'))
}
\keyword{internal}
