% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathervane.R
\name{earliest_dataset_date}
\alias{earliest_dataset_date}
\title{(Private) The earliest date of data available}
\usage{
earliest_dataset_date()
}
\value{
A Date object representing the earliest date
}
\description{
Return a Date object containing the earliest date of data
available from the SILO server. (As of checking on 10/03/2021,
the earliest date is 01/01/1889, although this may change.)
}
\examples{
weathervane:::earliest_dataset_date()
}
