% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_all_stations}
\alias{get_all_stations}
\title{Get list of all weather stations}
\usage{
get_all_stations(sort_by = "name")
}
\arguments{
\item{sort_by}{The column to sort the stations by. Valid values are "name" (the default), "id" or "state".}
}
\value{
A data.frame with all the weather stations along with their BoM station ID, Station name, Latitude, Longitude, State and Elevation.
}
\description{
Get the complete list of weather stations with data available on SILO (approximately 8000).
}
\examples{
head(get_all_stations())
head(get_all_stations(sort_by = "id"))

}
