% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_station_by_name}
\alias{get_station_by_name}
\title{Get station details from a provided (partial) name}
\usage{
get_station_by_name(station)
}
\arguments{
\item{station}{A (partial) name of a station. Will be truncated to 10 characters. Wildcard searching can be performed with * or _}
}
\value{
A data frame listing all stations that matched the input text.
}
\description{
Get station details from a provided (partial) name
}
\examples{
# Search by name
get_station_by_name("Brisbane")

# Find a specific station
get_station_by_name("Adel (Waite)")

# Will return any stations containing "botanic", "botanical" or "botany"
get_station_by_name("botan*")
}
