% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_stations_by_dist}
\alias{get_stations_by_dist}
\title{Get list of all weather stations within specified distance}
\usage{
get_stations_by_dist(station, distance, sort_by = "distance")
}
\arguments{
\item{station}{The station to retrieve details of. Station names will be attempted to be interpreted with an error returned.}

\item{distance}{Radius in km from provided station.}

\item{sort_by}{The column to sort the stations by. Valid values are "name" (the default), "distance", "id" or "state".}
}
\value{
A data.frame with all the weather stations along with their BoM station ID, Station name, Latitude, Longitude, State and Elevation.
}
\description{
Get a list of weather stations within a provided distance from a given weather station.
}
\examples{
get_stations_by_dist("Waite", 5)
get_stations_by_dist(23031, 5)

}
