% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_view.R
\name{VariableView}
\alias{VariableView}
\title{TODO: Making this as an R6 object, so document accordingly.}
\description{
TODO: Making this as an R6 object, so document accordingly.

TODO: Making this as an R6 object, so document accordingly.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariableView$new()}}
\item \href{#method-ui}{\code{VariableView$ui()}}
\item \href{#method-draw_plot}{\code{VariableView$draw_plot()}}
\item \href{#method-observers}{\code{VariableView$observers()}}
\item \href{#method-clone}{\code{VariableView$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableView$new(var, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableView$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-draw_plot"></a>}}
\if{latex}{\out{\hypertarget{method-draw_plot}{}}}
\subsection{Method \code{draw_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableView$draw_plot(output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-observers"></a>}}
\if{latex}{\out{\hypertarget{method-observers}{}}}
\subsection{Method \code{observers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableView$observers(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableView$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
