% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathervane.R
\name{weather_variables}
\alias{weather_variables}
\title{Weather variables, including ID codes and descriptions}
\usage{
weather_variables()
}
\value{
A data.frame containing the weather variables and codes
}
\description{
Return a data frame mapping the available weather variables to
their 'pretty' names, their descriptions, and their SILO
codes and identifier names.
(See also
\url{https://www.longpaddock.qld.gov.au/silo/about/climate-variables/})
}
\examples{
weather_variables()
}
