% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_station_details}
\alias{get_station_details}
\title{Get details of an individual weather station}
\usage{
get_station_details(station_id)
}
\arguments{
\item{station_id}{The station ID to retrieve details of.}
}
\value{
A data.frame with Station ID, Name, Latitude, Longitude, State and Elevation of the given station ID.
}
\description{
Get details of an individual weather station
}
\examples{
get_station_details(1001)
}
