% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_weather_app.R
\name{run_weather_app}
\alias{run_weather_app}
\title{Run the Shiny weathervane app}
\usage{
run_weather_app(browser = FALSE)
}
\arguments{
\item{browser}{A logical flag to run app directly in the browser}
}
\value{
A list containing ggplot2 objects which are diagnostic plots.
}
\description{
Runs a shiny app where a user can select and download weather data.
}
\examples{
\dontrun{
run_weather_app()
}
}
